Moves obs into plugins.obs
