+++ plugins/obs/static/js/obs/controllers/obs_chart.js
