+++ plugins/obs/static/js/obs/services/records/observation.js
