+++ plugins/obs/static/js/test/obs_chart.test.js
