+++ plugins/obs/templates/obs/chart.html
