+++ plugins/obs/templates/records/observation.html
